cargo: Sync up MSRV with CI by setting `rust-version`
This fixes the `+nightly -Zminimal-versions generate-lockfile` step
to no longer generate a `version = 4` `Cargo.lock` which is incompatible
with our 1.74 MSRV.

And it communicates to crates.io and newer `cargo` versions with
MSRV-aware resolver what our crate is going to be compatible with.

Also drop the `macos-12` runner CI: GitHub no longer hosts it, and we
already have `macos-13` and `macos-latest` anyway.