Pregenerate bindings for target platforms (#36)
Pregenerate the bindings through Generate bindings Github action workflow and check them into the code base. As a result, the bindgen crate can become optional again for crates that want to depend on this crate. For now, the following targets have pregenerated bindings:

    x86_64-unknown-linux-gnu
    x86_64-apple-darwin
    x86_64-pc-windows-msvc
    aarch64-apple-darwin
    aarch64-linux-android

More bindings can be added when desired.